+++ test/bench.js
@@ -1,13 +1,11 @@
 var CleanCSS = require('../index');
-var path = require('path');
 
-var benchDir = path.join(__dirname, 'fixtures', 'bench');
-var cssData = require('fs').readFileSync(path.join(benchDir, 'complex.css'), 'utf8');
+var input = '@import url(test/fixtures/bench/complex.css);';
 var total = 0;
 
 for (var i = 1; i <= 10; i++) {
   var start = process.hrtime();
-  new CleanCSS({ benchmark: i == 10, root: benchDir }).minify(cssData);
+  new CleanCSS({ benchmark: i == 10 }).minify(input);
 
   var itTook = process.hrtime(start);
   total += 1000 * itTook[0] + itTook[1] / 1000000;